modify all API calls to cl_dev_err_code instead of cl_uint

+++ opencl-intel/fe_compilers/clang_compiler/clang_compiler.cpp
@@ -62,7 +62,7 @@ extern "C" DLL_EXPORT	int	clFEBuildProgram(FEBuildProgramDesc* pDesc)
 
 	ITask* pTask = new CompileTask(pDesc);
 	if ( NULL == pTask )
-		return CL_DEV_OUT_OF_MEMORY;
+		return (int) CL_DEV_OUT_OF_MEMORY;
 
 	GetTaskExecutor()->Execute(pTask);
 	LOG_INFO(TEXT("%S"), TEXT("Exit"));
